Exception handling - Delete Update

+++ src/main/java/com/munninlabs/springboot_jpa/services/UserService.java
@@ -2,8 +2,12 @@ package com.munninlabs.springboot_jpa.services;
 
 import com.munninlabs.springboot_jpa.entities.User;
 import com.munninlabs.springboot_jpa.repositories.UserRepository;
+import com.munninlabs.springboot_jpa.services.exceptions.DatabaseException;
 import com.munninlabs.springboot_jpa.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -13,6 +17,9 @@ import java.util.Optional;
 public class UserService {
     @Autowired
     private UserRepository userRepository;
+
+    private static final org.slf4j.Logger log = org.slf4j.LoggerFactory.getLogger(UserService.class);
+
 
     public List<User> findAll() {
         return userRepository.findAll();
@@ -28,13 +35,25 @@ public class UserService {
     }
 
     public void deleteUser(Long id) {
-        userRepository.deleteById(id);
+        if (!userRepository.existsById(id)) {
+            throw new ResourceNotFoundException(id);
+        }
+        try {
+            userRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            log.error("Database integrity violation", e);
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User updateUser(Long Id, User user) {
-        User entity = userRepository.getReferenceById(Id);
-        updateData(entity, user);
-        return userRepository.save(entity);
+        try {
+            User entity = userRepository.getReferenceById(Id);
+            updateData(entity, user);
+            return userRepository.save(entity);
+        } catch ( EntityNotFoundException e ) {
+            throw new ResourceNotFoundException(Id);
+        }
     }
 
     private void updateData(User entity, User user) {
